@@ -18,7 +18,7 @@ publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0 <3.0.0'
 
 dependencies:
   flutter:
@@ -27,22 +27,22 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.0
-  get: ^3.26.0
-  firebase_core: '0.7.0'
-  cloud_firestore: '^0.16.0+1'
-  firebase_auth: '^0.20.1'
-  firebase_storage: '^7.0.0'
-  intl: ^0.16.1
-  timeago: ^2.0.30
-  carousel_slider: ^3.0.0
-  flutter_screenutil: ^4.0.3+1
-  dartz: ^0.9.2
-  equatable: ^1.2.6
+  get: ^4.1.4
+  firebase_core: '^1.2.1'
+  cloud_firestore: '^2.2.1'
+  firebase_auth: '^1.3.0'
+  firebase_storage: '^8.1.1'
+  intl: ^0.17.0
+  timeago: ^3.0.2
+  carousel_slider: ^4.0.0-nullsafety.0
+  flutter_screenutil: ^5.0.0+2
+  dartz: ^0.10.0-nullsafety.2
+  equatable: ^2.0.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  mockito: ^4.1.0
+  mockito: ^5.0.10
 
 # For information on the generic Dart part of this file, see the
 # following page: https://dart.dev/tools/pub/pubspec
