@@ -2,5 +2,5 @@ import 'package:flutter/foundation.dart';
 
 class ErrorModel {
   final String message;
-  ErrorModel({@required this.message});
+  ErrorModel({required this.message});
 }
