@@ -1,8 +1,8 @@
 class Core {
-  DateTime createdAt;
-  DateTime updatedAt;
-  bool isDeleted;
-  DateTime deletedAt;
+  DateTime? createdAt;
+  DateTime? updatedAt;
+  bool? isDeleted;
+  DateTime? deletedAt;
 
   Core(this.createdAt, this.updatedAt);
 }
