@@ -3,17 +3,17 @@ import 'package:dream/models/core.dart';
 import 'package:flutter/foundation.dart';
 
 class PrayModel extends Core {
-  String id;
-  String userId;
-  String title;
-  String content;
-  List<String> prayUserList;
-  DocumentReference documentReference;
+  String? id;
+  String? userId;
+  String? title;
+  String? content;
+  List<String>? prayUserList;
+  DocumentReference? documentReference;
 
   PrayModel(
-      {@required this.userId,
-      @required this.title,
-      @required this.content,
+      {required this.userId,
+      required this.title,
+      required this.content,
       this.prayUserList = const []})
       : super(DateTime.now(), DateTime.now());
 
@@ -27,8 +27,8 @@ class PrayModel extends Core {
     );
     model.id = documentSnapshot.id;
     model.documentReference = documentSnapshot.reference;
-    model.createdAt = (json['created_at'] as Timestamp)?.toDate() ?? null;
-    model.updatedAt = (json['updated_at'] as Timestamp)?.toDate() ?? null;
+    model.createdAt = (json['created_at'] as Timestamp?)?.toDate() ?? null;
+    model.updatedAt = (json['updated_at'] as Timestamp?)?.toDate() ?? null;
     return model;
   }
 
