@@ -4,20 +4,20 @@ import 'package:equatable/equatable.dart';
 import 'package:flutter/foundation.dart';
 
 class ReplyModel extends Core with EquatableMixin {
-  final String id;
-  final String userId;
-  final String content;
-  List<String> favoriteUserList;
+  final String? id;
+  final String? userId;
+  final String? content;
+  List<String>? favoriteUserList;
 
   ReplyModel(
-      {@required this.id,
-      @required this.userId,
-      @required this.content,
-      @required this.favoriteUserList})
+      {required this.id,
+      required this.userId,
+      required this.content,
+      required this.favoriteUserList})
       : super(DateTime.now(), DateTime.now());
 
   @override
-  List<Object> get props => [userId, content, favoriteUserList];
+  List<Object?> get props => [userId, content, favoriteUserList];
 
   factory ReplyModel.fromJson(Map<String, dynamic> json) {
     var model = ReplyModel(
@@ -26,8 +26,8 @@ class ReplyModel extends Core with EquatableMixin {
       content: json['content'],
       favoriteUserList: json['favorite_user_list']?.cast<String>()?.toList(),
     );
-    model.createdAt = (json['created_at'] as Timestamp)?.toDate() ?? null;
-    model.updatedAt = (json['updated_at'] as Timestamp)?.toDate() ?? null;
+    model.createdAt = (json['created_at'] as Timestamp?)?.toDate() ?? null;
+    model.updatedAt = (json['updated_at'] as Timestamp?)?.toDate() ?? null;
     return model;
   }
 
