@@ -11,19 +11,19 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'image_spliter.dart';
 
 class NoticeCard extends StatelessWidget {
-  final NoticeModel notice;
+  final NoticeModel? notice;
   final bool isCommentScreen;
   //TODO: 실제 유저 아이디로 변경해야함.
   final String userId = '123';
 
   const NoticeCard(
-      {Key key, @required this.notice, this.isCommentScreen = false})
+      {Key? key, required this.notice, this.isCommentScreen = false})
       : super(key: key);
   @override
   Widget build(BuildContext context) {
     var size = MediaQuery.of(context).size;
     //시간 로드
-    var date = TimeUtil.getDateString(notice.updatedAt ?? notice.createdAt);
+    var date = TimeUtil.getDateString(notice!.updatedAt ?? notice!.createdAt!);
     NoticeViewModel noticeViewModel = Get.find<NoticeViewModel>();
 
     return Padding(
@@ -43,24 +43,24 @@ class NoticeCard extends StatelessWidget {
                   SizedBox(
                     height: 5.h,
                   ),
-                  Text(notice.content,
+                  Text(notice!.content!,
                       style: Theme.of(context).textTheme.bodyText1),
                   SizedBox(
                     height: 8.h,
                   ),
                   //이미지, 슬라이더
-                  if (notice.imageList.length > 0)
+                  if (notice!.imageList.length > 0)
                     GestureDetector(
                         onTap: () {
                           // 화면 전환
                           Get.to(() => ImageSliderScreen(
-                                imageUrlList: notice.imageList,
+                                imageUrlList: notice!.imageList,
                               ));
                         },
                         child: Container(
                           width: size.width,
                           height: 250.w,
-                          child: ImageSpliter(images: notice.imageList),
+                          child: ImageSpliter(images: notice!.imageList),
                         )),
                 ],
               ),
@@ -76,7 +76,7 @@ class NoticeCard extends StatelessWidget {
                   GestureDetector(
                     onTap: () {
                       noticeViewModel.toggleNoticeFavorite(
-                          noticeId: notice.id, userId: userId);
+                          noticeId: notice!.id, userId: userId);
                       if (isCommentScreen) {
                         Get.find<CommentReplyViewModel>().refreshComment();
                       }
@@ -84,7 +84,7 @@ class NoticeCard extends StatelessWidget {
                     behavior: HitTestBehavior.opaque,
                     child: Row(
                       children: [
-                        if (notice.favoriteUserList
+                        if (notice!.favoriteUserList!
                             .where((u) => u == userId)
                             .isNotEmpty)
                           Icon(
@@ -99,7 +99,7 @@ class NoticeCard extends StatelessWidget {
                         Container(
                           width: 5.w,
                         ),
-                        Text("좋아요 ${notice.favoriteUserList.length ?? ""}",
+                        Text("좋아요 ${notice!.favoriteUserList!.length ?? ""}",
                             style: Theme.of(context).textTheme.bodyText2)
                       ],
                     ),
@@ -116,7 +116,7 @@ class NoticeCard extends StatelessWidget {
                       ),
                       Text(
                           //TODO: NULL처리 해줘야함.
-                          '댓글 ${notice.commentCount < 1 ? '쓰기' : notice.commentCount}',
+                          '댓글 ${notice!.commentCount! < 1 ? '쓰기' : notice!.commentCount}',
                           style: Theme.of(context).textTheme.bodyText2)
                     ],
                   )
