@@ -8,25 +8,25 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 import 'package:get/get.dart';
 
 class NoticeReply extends StatelessWidget {
-  final String noticeId;
-  final String commentId;
+  final String? noticeId;
+  final String? commentId;
   final ReplyModel replyModel;
-  final Function deleteReply;
+  final Function? deleteReply;
   CommentReplyViewModel commentReplyViewModel =
       Get.find<CommentReplyViewModel>();
   NoticeReply({
-    Key key,
-    @required this.noticeId,
-    @required this.commentId,
-    @required this.replyModel,
-    @required this.deleteReply,
+    Key? key,
+    required this.noticeId,
+    required this.commentId,
+    required this.replyModel,
+    required this.deleteReply,
   }) : super(key: key);
 
   @override
   Widget build(BuildContext context) {
     //시간 로드
     var date =
-        TimeUtil.getDateString(replyModel.updatedAt ?? replyModel.createdAt);
+        TimeUtil.getDateString(replyModel.updatedAt ?? replyModel.createdAt!);
 
     return Container(
       color: Colors.white,
@@ -60,7 +60,7 @@ class NoticeReply extends StatelessWidget {
                               padding: const EdgeInsets.symmetric(vertical: 5),
                               child: Text(
                                 //TODO: 추후에 닉네임으로 변경해야함.
-                                replyModel.userId,
+                                replyModel.userId!,
                                 style: TextStyle(fontWeight: FontWeight.bold),
                               ),
                             ),
@@ -77,7 +77,7 @@ class NoticeReply extends StatelessWidget {
                           padding: const EdgeInsets.only(top: 8.0),
                           child: InkWell(
                               onTap: () {
-                                deleteReply(commentId, replyModel);
+                                deleteReply!(commentId, replyModel);
                               },
                               child: Icon(Icons.more_vert)),
                         ),
@@ -89,7 +89,7 @@ class NoticeReply extends StatelessWidget {
                     child: Container(
                       width: 180.w,
                       child: Text(
-                        replyModel.content,
+                        replyModel.content!,
                         overflow: TextOverflow.clip,
                       ),
                     ),
@@ -123,7 +123,7 @@ class NoticeReply extends StatelessWidget {
                         child: Row(
                           children: [
                             //TODO: 가짜 유저 아이디 실제 유저 아이디로 변경 필요
-                            if (replyModel.favoriteUserList
+                            if (replyModel.favoriteUserList!
                                 .where((userId) => userId == '123')
                                 .isNotEmpty)
                               Icon(Icons.favorite)
@@ -132,7 +132,7 @@ class NoticeReply extends StatelessWidget {
                             SizedBox(
                               width: 5,
                             ),
-                            Text("${replyModel.favoriteUserList.length ?? "0"}")
+                            Text("${replyModel.favoriteUserList!.length ?? "0"}")
                           ],
                         ),
                       )
