@@ -8,20 +8,20 @@ import 'package:flutter/foundation.dart';
 import 'package:dream/constants.dart';
 
 class ReplyRepositoryImpl extends ReplyRepository {
-  FirebaseFirestore _firebaseFirestore;
+  late FirebaseFirestore _firebaseFirestore;
   // FirebaseStorage _firebaseStorage = FirebaseStorage.instance;
 
-  ReplyRepositoryImpl({@required FirebaseFirestore firebaseFirestore}) {
+  ReplyRepositoryImpl({required FirebaseFirestore firebaseFirestore}) {
     _firebaseFirestore = firebaseFirestore;
   }
 
   @override
   Future<Either<ErrorModel, void>> writeReply(
-      {@required String noticeId,
-      @required String commentId,
-      @required String replyIndex,
-      @required String userId,
-      @required String content}) async {
+      {required String? noticeId,
+      required String? commentId,
+      required String replyIndex,
+      required String userId,
+      required String content}) async {
     try {
       //TODO: Notice Reference 혹은 comment Reference 필요
       var replyModel = ReplyModel(
@@ -31,9 +31,9 @@ class ReplyRepositoryImpl extends ReplyRepository {
           favoriteUserList: []);
       await _firebaseFirestore
           .collection(noticeCollectionName)
-          .doc(noticeId)
+          .doc(noticeId!)
           .collection(commentCollectionName)
-          .doc(commentId)
+          .doc(commentId!)
           .update({
         'reply_list': FieldValue.arrayUnion([replyModel.toJson()])
       });
@@ -45,14 +45,14 @@ class ReplyRepositoryImpl extends ReplyRepository {
 
   @override
   Future<Either<ErrorModel, void>> deleteReply({
-    @required String noticeId,
-    @required String commentId,
-    @required ReplyModel replyModel,
+    required String? noticeId,
+    required String commentId,
+    required ReplyModel replyModel,
   }) async {
     try {
       await _firebaseFirestore
           .collection(noticeCollectionName)
-          .doc(noticeId)
+          .doc(noticeId!)
           .collection(commentCollectionName)
           .doc(commentId)
           .update({
@@ -66,25 +66,25 @@ class ReplyRepositoryImpl extends ReplyRepository {
 
   @override
   Future<Either<ErrorModel, void>> toggleReplyFavorite({
-    @required String noticeId,
-    @required String commentId,
-    @required ReplyModel reply,
-    @required String userId,
+    required String? noticeId,
+    required String? commentId,
+    required ReplyModel reply,
+    required String userId,
   }) async {
     try {
       await _firebaseFirestore
           .collection(noticeCollectionName)
-          .doc(noticeId)
+          .doc(noticeId!)
           .collection(commentCollectionName)
-          .doc(commentId)
+          .doc(commentId!)
           .update({
         'reply_list': FieldValue.arrayRemove([reply.toJson()])
       });
       //이미 유저 있다면 삭제 없다면 추가
-      if (reply.favoriteUserList.where((e) => e == userId).isEmpty)
-        reply.favoriteUserList.add(userId);
+      if (reply.favoriteUserList!.where((e) => e == userId).isEmpty)
+        reply.favoriteUserList!.add(userId);
       else
-        reply.favoriteUserList.remove(userId);
+        reply.favoriteUserList!.remove(userId);
 
       var modifyReply = reply.toJson();
 
