@@ -3,7 +3,7 @@ import 'package:flutter_screenutil/flutter_screenutil.dart';
 
 class Constants {
   static Color backgroundColor = Colors.white;
-  static Color primaryColor = Colors.amber[300];
+  static Color? primaryColor = Colors.amber[300];
   static Color iconColor = Colors.black26;
   static Color iconSelectedColor = Colors.black;
   static Color favoriteAndCommentColor = Colors.black;
